requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
deepface==0.0.93
easyocr==1.7.1
filelock==3.16.1
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers @ file:///croot/python-flatbuffers_1722369094438/work
fonttools==4.53.1
fsspec==2024.9.0
ftfy==6.2.3
gast @ file:///tmp/build/80754af9/gast_1637837472536/work
gdown==5.2.0
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///home/conda/feedstock_root/build_artifacts/grpc-split_1713388282847/work
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1717664837666/work
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
itsdangerous==2.2.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
Markdown @ file:///work/ci_py311/markdown_1676837083112/work
markdown-it-py @ file:///croot/markdown-it-py_1684279902645/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mdurl @ file:///work/ci_py311/mdurl_1676827289376/work
mkl-fft==1.3.1
mkl-random @ file:///work/ci_py311/mkl_random_1676828032909/work
mkl-service==2.4.0
ml_dtypes @ file:///croot/ml_dtypes_1722369114743/work
mpmath @ file:///croot/mpmath_1690848262763/work
mtcnn==0.1.1
namex @ file:///croot/namex_1715206936162/work
networkx @ file:///croot/networkx_1720002482208/work
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
ollama==0.3.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
optree @ file:///croot/optree_1723069484006/work
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.2
patchify==0.2.3
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1712154447422/work
priority==2.0.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.1.4
pypdfium2==4.30.0
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
qrdet==2.5
qreader==3.14
quadrilateral-fitter==1.12
Quart==0.19.6
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
retina-face==0.0.17
rich @ file:///croot/rich_1720637495510/work
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.6
surya-ocr==0.5.0
sympy @ file:///croot/sympy_1724938189289/work
tabulate==0.9.0
tensorboard @ file:///croot/tensorboard_1723042039441/work/tensorboard-2.17.0-py3-none-any.whl#sha256=859a499a9b1fb68a058858964486627100b71fcb21646861c61d31846a6478fb
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=c166a9f297fa12121b89b94be4a4cc150aa7996c9598410ba057ecd592eea8c7
tensorflow @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1725105520986/work/tensorflow_pkg/tensorflow-2.17.0-cp311-cp311-linux_x86_64.whl#sha256=ffaa4ee6f243d7e7362107bb01ef027dc786b425a6d720cc2adcfbc844b5b609
tensorflow_estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1725105520986/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.15.0-py2.py3-none-any.whl#sha256=d5de76e226ccdc61f4bec1e708b3e897a25f62b90ceab7a048e5511582785d71
termcolor @ file:///work/ci_py311/termcolor_1676838860478/work
tf_keras==2.17.0
tifffile==2024.8.30
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ultralytics==8.2.95
ultralytics-thop==2.0.6
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.6
wcwidth==0.2.13
Werkzeug @ file:///croot/werkzeug_1716533313107/work
wrapt @ file:///work/ci_py311/wrapt_1676827586170/work
wsproto==1.2.0
